YSY/mic_say/requirements.txt
SpeechRecognition
gTTS
pygame
python-dotenv
rich
google-generativeai
